requirements.txt
streamlit
spacy
pandas
python-docx
pdfplumber
PyPDF2
tqdm
